refactor: Make CustomDrugForm separate component
